TEST-#2288: addressing review comments
